+++ packages/remix/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 4.13.19
+
+### Patch Changes
+
+- Updated dependencies [[`d8f59a6`](https://github.com/clerk/javascript/commit/d8f59a66d56d8fb0dfea353ecd86af97d0ec56b7)]:
+  - @clerk/shared@3.35.2
+  - @clerk/backend@2.23.2
+  - @clerk/clerk-react@5.56.2
+  - @clerk/types@4.101.2
+
 ## 4.13.18
 
 ### Patch Changes